Restructured fpaths.
